+++ CodeChat_Server/tests/test_server.py
@@ -1,6 +1,7 @@
 # ****************************************************
 # |docname| - Tests for `../CodeChat_Server/server.py`
 # ****************************************************
+# TODO: create a CodeChat client. Use it to test the server. Import the event to shut down the server when needed. Run the server in a different thread from the fixture; drop all the subprocess stuff.
 #
 # Imports
 # =======
@@ -9,14 +10,28 @@
 # Standard library
 # ----------------
 import socketserver
+import subprocess
+import sys
 
 # Third-party imports
 # -------------------
-# None.
-#
+import pytest
+
 # Local imports
 # -------------
 from CodeChat_Server.server import is_port_in_use
+
+
+# Fixtures
+# ========
+@pytest.fixture
+def run_servers():
+    p = subprocess.Popen(
+        [sys.argv[0], "-m", "CodeChat_Server"], stdin=subprocess.PIPE, text=True
+    )
+    yield
+    p.communicate("q\n")
+    p.wait()
 
 
 def test_1():